refactor(native-rd): delete ThemeChipGrid, now that nothing mounts it
Replacing it with ThemeSwatchRail on WelcomeScreen removed its last production
importer — only its own story and tests referenced it. CLAUDE.md says delete
unreachable code, and leaving it shipped a second theme picker still carrying
the #500 a11y bug (accessible={true} outside E2E collapses all 7 options into
one VoiceOver node) with no live surface to fix it on.

Its tests are all mirrored by ThemeSwatchRail's: radiogroup role, one labelled
radio per option, single checked option, select-by-id, and the E2E-gating
regression — which the rail covers in its fixed form. Nothing unique is lost.

swatch-utils.ts moves to ThemeSwatchRail/, its only remaining consumer. Also
refreshed two stale comments that pointed at the deleted component.
