--- apps/native-rd/src/components/ThemeChipGrid/index.ts
@@ -1,1 +1,0 @@
-export { ThemeChipGrid } from "./ThemeChipGrid";
